services/auth/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
dependencies = [
  "fastapi>=0.112.1",
  "pydantic[email]>=2.8.2",
  "pydantic-settings>=2.4.0",
  "fastapi-pagination>=0.12.26",
  "backoff>=2.2.1",
  "uvicorn>=0.30.6",
  "gunicorn>=23.0.0",
  "uvloop>=0.20.0",
  "orjson>=3.10.7",
  "SQLAlchemy>=2.0.29",
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "psycopg2-binary>=2.9.9",
  "httpx>=0.27.0",
  "async-timeout>=4.0.3",
  "redis>=5.0.8",
  "werkzeug>=3.0.4",
  "python-multipart>=0.0.9",
  "python-jose>=3.3.0",
  "cryptography>=43.0.0",
  "requests>=2.32.3",
  "fastapi-limiter>=0.1.6",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-exporter-jaeger>=1.21.0",
  "sentry-sdk[fastapi]>=2.13.0",
  "python-dotenv>=1.0.1",
]
name = "Auth"
requires-python = ">=3.12"
version = "0.3.0"

[project.optional-dependencies]
test = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "aiohttp>=3.10.5",
]
test-functional = []
test-integration = [
  "requests>=2.32.3",
]

[tool.hatch.build.targets.wheel]
packages = ["services/auth"]

[tool.pytest.ini_options]
addopts = "-vv -p no:cacheprovider"
log_cli = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_level = "DEBUG"
